requirements.txt
altair                    5.2.0
attrs                     23.2.0
blinker                   1.7.0
cachetools                5.3.2
certifi                   2023.11.17
charset-normalizer        3.3.2
click                     8.1.7
colorama                  0.4.6
gitdb                     4.0.11
GitPython                 3.1.41
idna                      3.6
importlib-metadata        7.0.1
Jinja2                    3.1.3
jsonschema                4.21.1
jsonschema-specifications 2023.12.1
markdown-it-py            3.0.0
MarkupSafe                2.1.4
mdurl                     0.1.2
numpy                     1.26.3
packaging                 23.2
pandas                    2.2.0
pillow                    10.2.0
pip                       23.3.2
protobuf                  4.25.2
pyarrow                   15.0.0
pydeck                    0.8.1b0
Pygments                  2.17.2
python-dateutil           2.8.2
pytz                      2023.3.post1
referencing               0.32.1
requests                  2.31.0
rich                      13.7.0
rpds-py                   0.17.1
six                       1.16.0
smmap                     5.0.1
streamlit                 1.30.0
tenacity                  8.2.3
toml                      0.10.2
toolz                     0.12.1
tornado                   6.4
typing_extensions         4.9.0
tzdata                    2023.4
tzlocal                   5.2
urllib3                   2.1.0
validators                0.22.0
watchdog                  3.0.0
zipp                      3.17.0
google-generativeai
streamlit
